More testing hax

+++ launch_chffrplus.sh
@@ -66,7 +66,8 @@ function launch {
 
   # Check for NEOS update
   # FIXME: hacked for testing, will force "update" if we have any NEOS files present
-  if [ $(< /VERSION) != "$REQUIRED_NEOS_VERSION"  ] || [ -d "/data/neoupdate" ]; then
+  if [ $(< /VERSION) != "$REQUIRED_NEOS_VERSION"  ] || ( [ -d "/data/neoupdate" ] && [ ! -f /data/neoupdate/ranonce ] ); then
+    touch /data/neoupdate/ranonce
     if [ -f "$BASEDIR/scripts/continue.sh" ]; then
       cp "$BASEDIR/scripts/continue.sh" "/data/data/com.termux/files/continue.sh"
     fi
